remove menu bar and switch to fullscreen as well
this remembers the previous menu bar setting, but not the full screen
setting.

we therefore assume full screen mode is always off before turning on
darkroom, which may be inaccurate.

we do our best with fullscreen, yet it doesn't always seem to be enough, see:

http://www.emacswiki.org/emacs/FullScreen

... for more information. it may also be useful to consider more frame
parameters, see:

https://www.gnu.org/software/emacs/manual/html_node/elisp/Management-Parameters.html#Management-Parameters

for more ideas. "sticky" strikes me as a possibly good idea.